refactor: move popup component code to files within a folder
